projects/04_enterprise_finops_dashboard/streamlit_app/requirements.txt
streamlit>=1.30.0
snowflake-snowpark-python>=1.11.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0
